Move script utils and types into seperate file
