+++ packages/semantic-tokens/package.json
@@ -42,7 +42,7 @@
     "lib-commonjs"
   ],
   "scripts": {
-    "generate-tokens": "npx ts-node scripts/legacyTokens.ts && npx ts-node scripts/tokenGen.ts",
+    "generate-tokens": "yarn generate-legacy-tokens && ts-node scripts/tokenGen.ts",
     "generate-legacy-tokens": "npx ts-node scripts/legacyTokens.ts"
   }
 }
